Split, namespace and reorganize core modules.
