+++ src/lua/cfg-core/strings.lua
@@ -1,0 +1,7 @@
+return {
+        IDENT = "Configi",
+        ERROR = "ERROR: ",
+         WARN = "WARNING: ",
+         SERR = "POLICY ERROR: ",
+    OPERATION = "Operation"
+}
