+++ src/lua/cfg-modules/sha256.lua
