+++ src/lua/cfg-modules/shell.lua
